+++ source/Public/Add-NTFSAccessRule.ps1
@@ -92,6 +92,7 @@ function Add-NTFSAccessRule {
         [string[]]$Account,
 
         [Parameter(Mandatory)]
+        [ArgumentCompleter([WindowsFileSystemRightsCompleter])]
         [WindowsAccessRightsTransformAttribute([System.Security.AccessControl.FileSystemRights])]
         $AccessRights,