+++ source/Public/Add-NTFSAuditRule.ps1
@@ -82,6 +82,7 @@ function Add-NTFSAuditRule {
         [string[]]$Account,
 
         [Parameter(Mandatory)]
+        [ArgumentCompleter([WindowsFileSystemRightsCompleter])]
         [WindowsAccessRightsTransformAttribute([System.Security.AccessControl.FileSystemRights])]
         $AccessRights,