+++ source/Public/Set-ADObjectAccessRule.ps1
@@ -79,6 +79,7 @@ function Set-ADObjectAccessRule {
         [Alias('IdentityReference', 'ID')]
         [object[]]$Account,
         [Parameter(Mandatory)]
+        [ArgumentCompleter([WindowsActiveDirectoryRightsCompleter])]
         [WindowsAccessRightsTransformAttribute([WindowsActiveDirectoryRights])]
         $AccessRights,
         [Parameter()]