+++ source/Public/Set-NTFSAuditRule.ps1
@@ -78,6 +78,7 @@ function Set-NTFSAuditRule {
         [string]$Account,
 
         [Parameter(Mandatory)]
+        [ArgumentCompleter([WindowsFileSystemRightsCompleter])]
         [WindowsAccessRightsTransformAttribute([System.Security.AccessControl.FileSystemRights])]
         $AccessRights,